requirements.txt
Pillow==9.5.0
geopy==2.3.0
